+++ app/(tabs)/_layout.tsx
@@ -1,0 +1,9 @@
+import { Tabs } from "expo-router";
+
+export default () => {
+  return (
+    <Tabs>
+      <Tabs.Screen name="buttons" />
+    </Tabs>
+  );
+};
