+++ app/_layout.tsx
@@ -1,0 +1,9 @@
+import { Stack } from "expo-router";
+
+export const StackLayout = () => {
+  return (
+    <Stack>
+      <Stack.Screen name="(tabs)" />
+    </Stack>
+  );
+};
