+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "esturide-client",
   "version": "0.0.1",
-  "main": "node_modules/expo/AppEntry.js",
+  "main": "expo-router/entry",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
@@ -21,11 +21,14 @@
     "@react-native-community/slider": "4.5.2",
     "expo": "^51.0.17",
     "expo-constants": "~16.0.2",
+    "expo-linking": "~6.3.1",
+    "expo-router": "~3.5.17",
     "expo-status-bar": "~1.12.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-native": "0.74.2",
     "react-native-safe-area-context": "4.10.1",
+    "react-native-screens": "3.31.1",
     "react-native-web": "~0.19.10"
   },
   "devDependencies": {
